requirements.txt
Django==2.1.7
djangorestframework==3.9.0
Pillow==5.3.0
pytz==2018.7
django-cors-headers==2.4.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
django-discover-runner==1.0
django-role-permissions==2.2.1
six==1.12.0
casl-django==0.4.2
django-import-export==1.1.0
drf-yasg==1.12.1
celery==4.2.1
pytg==0.4.10
django-scheduler==0.8.8
django-cron==0.5.1
django-cities-light==3.5.0
mysqlclient==1.4.2.post1